dev-requirements.txt
aiohttp==1.0.5
appnope==0.1.0
astroid==1.4.9
async-timeout==1.1.0
chardet==2.3.0
decorator==4.0.11
discord.py==0.16.5
ipdb==0.10.2
ipython==5.2.2
ipython-genutils==0.1.0
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.6.1
multidict==2.1.4
pexpect==4.2.1
pickleshare==0.7.4
praw==4.3.0
prawcore==0.7.0
prompt-toolkit==1.0.13
ptyprocess==0.5.1
pyenchant==1.6.8
pygeoip==0.3.2
Pygments==2.2.0
pylint==1.6.5
pytz==2016.10
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
sopel==6.5.0
traitlets==4.3.1
update-checker==0.16
wcwidth==0.1.7
websockets==3.2
wrapt==1.10.8
xmltodict==0.10.2
